+++ CHANGELOG.md
@@ -61,6 +61,7 @@ All notable changes to this project will be documented in this file.
 - Added ratio size. This size is relative to its parent size.
 - Added `plugin` field to Outputs modal and command
 - Fixed order when adding multiple items with `Insert` and `InsertAll`
+- Added `Position` property which shows song's position in the Queue
 
 ### Changed